Revert "refactor(bindings): move error logic to errors module (#1076)"
This reverts commit 60519770b14258a06bc53454f98e0e8589a1390a.
