+++ data-plane/bindings/adapter/src/session_context.rs
@@ -13,9 +13,9 @@ use slim_datapath::messages::utils::{PUBLISH_TO, SlimHeaderFlags, TRUE_VAL};
 use slim_session::SessionError;
 use slim_session::context::SessionContext;
 
-use crate::adapter::{FfiCompletionHandle, ReceivedMessage, SessionType};
+use crate::Name;
+use crate::adapter::{FfiCompletionHandle, ReceivedMessage, SessionType, SlimError};
 use crate::message_context::MessageContext;
-use crate::{Name, SlimError};
 
 /// Session context for language bindings (UniFFI-compatible)
 ///
@@ -586,7 +586,6 @@ impl BindingsSessionContext {
 mod tests {
     use super::*;
     use crate::Name as FfiName;
-    use crate::errors::SlimError;
     use slim_datapath::api::{
         ApplicationPayload, ProtoMessage, ProtoPublish, ProtoPublishType, SessionHeader, SlimHeader,
     };
@@ -916,7 +915,7 @@ mod tests {
         let result = ctx.publish_async(b"test".to_vec(), None, None).await;
         assert!(result.is_err());
         match result.unwrap_err() {
-            SlimError::SessionError { message } => {
+            crate::adapter::SlimError::SessionError { message } => {
                 assert!(message.contains("closed") || message.contains("dropped"));
             }
             _ => panic!("Expected SessionError"),
@@ -931,7 +930,7 @@ mod tests {
             .await;
         assert!(result.is_err());
         match result.unwrap_err() {
-            SlimError::SessionError { message } => {
+            crate::adapter::SlimError::SessionError { message } => {
                 assert!(message.contains("closed") || message.contains("dropped"));
             }
             _ => panic!("Expected SessionError"),
@@ -954,7 +953,7 @@ mod tests {
             .publish_to_async(message_ctx, b"reply".to_vec(), None, None)
             .await;
         assert!(result.is_err_and(|e| {
-            if let SlimError::SessionError { message } = e {
+            if let crate::adapter::SlimError::SessionError { message } = e {
                 message.contains("closed")
             } else {
                 false
@@ -978,7 +977,7 @@ mod tests {
             .publish_to_with_completion_async(message_ctx, b"reply".to_vec(), None, None)
             .await;
         assert!(result.is_err_and(|e| {
-            if let SlimError::SessionError { message } = e {
+            if let crate::adapter::SlimError::SessionError { message } = e {
                 message.contains("closed")
             } else {
                 false
@@ -1000,7 +999,7 @@ mod tests {
             .publish_with_params_async(dest, 1, b"test".to_vec(), None, None, None)
             .await;
         assert!(result.is_err_and(|e| {
-            if let SlimError::SessionError { message } = e {
+            if let crate::adapter::SlimError::SessionError { message } = e {
                 message.contains("closed")
             } else {
                 false
@@ -1068,7 +1067,7 @@ mod tests {
 
         let result = ctx.get_message_async(Some(10)).await;
         assert!(result.is_err_and(|e| {
-            if let SlimError::SessionError { message } = e {
+            if let crate::adapter::SlimError::SessionError { message } = e {
                 message.contains("receive timeout")
             } else {
                 false
@@ -1082,7 +1081,7 @@ mod tests {
         drop(tx);
 
         let result = ctx.get_message_async(Some(100)).await;
-        assert!(result.is_err_and(|e| matches!(e, SlimError::SessionError { .. })));
+        assert!(result.is_err_and(|e| matches!(e, crate::adapter::SlimError::SessionError { .. })));
     }
 
     // ==================== Invite/Remove FFI Tests ====================
@@ -1099,7 +1098,7 @@ mod tests {
 
         let result = ctx.invite_async(participant).await;
         assert!(result.is_err_and(|e| {
-            if let SlimError::SessionError { message } = e {
+            if let crate::adapter::SlimError::SessionError { message } = e {
                 message.contains("closed")
             } else {
                 false
@@ -1119,7 +1118,7 @@ mod tests {
 
         let result = ctx.remove_async(participant).await;
         assert!(result.is_err_and(|e| {
-            if let SlimError::SessionError { message } = e {
+            if let crate::adapter::SlimError::SessionError { message } = e {
                 message.contains("closed")
             } else {
                 false
@@ -1135,7 +1134,7 @@ mod tests {
         let result = ctx.destination();
         assert!(result.is_err());
         match result.unwrap_err() {
-            SlimError::SessionError { message } => {
+            crate::adapter::SlimError::SessionError { message } => {
                 assert!(message.contains("closed") || message.contains("dropped"));
             }
             _ => panic!("Expected SessionError"),
@@ -1148,7 +1147,7 @@ mod tests {
         let result = ctx.source();
         assert!(result.is_err());
         match result.unwrap_err() {
-            SlimError::SessionError { message } => {
+            crate::adapter::SlimError::SessionError { message } => {
                 assert!(message.contains("closed") || message.contains("dropped"));
             }
             _ => panic!("Expected SessionError"),
@@ -1161,7 +1160,7 @@ mod tests {
         let result = ctx.session_id();
         assert!(result.is_err());
         match result.unwrap_err() {
-            SlimError::SessionError { message } => {
+            crate::adapter::SlimError::SessionError { message } => {
                 assert!(message.contains("closed") || message.contains("dropped"));
             }
             _ => panic!("Expected SessionError"),
@@ -1174,7 +1173,7 @@ mod tests {
         let result = ctx.session_type();
         assert!(result.is_err());
         match result.unwrap_err() {
-            SlimError::SessionError { message } => {
+            crate::adapter::SlimError::SessionError { message } => {
                 assert!(message.contains("closed") || message.contains("dropped"));
             }
             _ => panic!("Expected SessionError"),
@@ -1187,7 +1186,7 @@ mod tests {
         let result = ctx.is_initiator();
         assert!(result.is_err());
         match result.unwrap_err() {
-            SlimError::SessionError { message } => {
+            crate::adapter::SlimError::SessionError { message } => {
                 assert!(message.contains("closed") || message.contains("dropped"));
             }
             _ => panic!("Expected SessionError"),
